moved BottomHalfH into landing-page
