+++ frontend/src/Pages/landing-page/components/shared/HeaderText.tsx
