+++ frontend/src/Pages/landing-page/components/shared/types.ts
