+++ frontend/src/Pages/landing-page/index.tsx
@@ -3,9 +3,9 @@ import Navbar from './components/Navbar'
 import HeroSection from './components/HeroSection'
 import FeaturesSection from './components/FeaturesSection'
 import AboutSection from './components/AboutSection'
-import Section_1 from '../../components/BottomHalfH/Section_1'
-import Section_2 from '../../components/BottomHalfH/Section_2'
-import Section_3 from '../../components/BottomHalfH/Section_3'
+import Section_1 from './components/BottomHalfH/Section_1'
+import Section_2 from './components/BottomHalfH/Section_2'
+import Section_3 from './components/BottomHalfH/Section_3'
 
 function LandingPage() {
   return (
